requirements.txt
alembic==1.9.4
bcrypt==4.0.1
black==22.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
docopt==0.6.2
earthengine-api==0.1.320
email-validator==1.3.1
et-xmlfile==1.1.0
Flask==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flit_core==3.6.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==1.12.11
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.21.0
httplib2shim==0.0.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
numpy==1.21.5
openpyxl==3.1.1
pandas==1.5.3
pathspec==0.10.3
pip==22.3.1
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==4.22.0
psycopg2==2.9.5
PuLP==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
setuptools==65.6.3
six==1.16.0
SQLAlchemy==1.4.20
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
wheel==0.38.4
wincertstore==0.2
WTForms==3.0.1
yarg==0.1.9
zipp==3.15.0
